# src/main/resources/api.yml
openapi: '3.0.0'
info:
  version: '1.0.0'
  title: 'Sample API'
  description: Initial example to submit Elasticsearch queries

paths:
  /search/fav:
    get:
      summary: 'Retrieves the favorite list'
      operationId: searchFavorites
      tags:
        - searchFavorites
      parameters:
        - name: pageNumber
          description: Number of current page
          in: query
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/ResultList'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /search/fav/{id}:
    post:
      summary: 'Add a document to favorite list'
      operationId: addToFavorites
      tags:
        - addToFavorites
      parameters:
        - name: id
          description: Document ID to be added to favorite list
          required: true
          in: path
          schema:
            type: string
      responses:
        201:
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: 'Document with ID {id} has been added to favorite list'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeFromFavorites
      tags:
        - removeFromFavorites
      summary: 'Remove a document from favorite list'
      parameters:
        - name: id
          description: Document ID to be removed from favorite list
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: 'Document with ID {id} has been removed from favorite list'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /search:
    get:
      summary: 'Submits a query to Elasticsearch'
      operationId: search
      tags:
        - search
      parameters:
        - name: queryParameter
          description: Object query to be submitted
          in: query
          schema:
              type: object
              properties:
                query:
                  type: string
                pageNumber:
                  type: integer
                filterReadingTime:
                  type: boolean
                  default: false
                filterDateCreation:
                  type: boolean
                  default: false
                filterMinValue:
                  type: integer
                  default: 0
                filterMaxValue:
                  type: integer
                  default: 999999
                filterMinDate:
                  type: string
                  default: '1970-01-01'
                filterMaxDate:
                  type: string
                  default: '9999-12-31'
                sortByTitle:
                  type: string
                sortByReadingTime:
                  type: string
                sortByDateCreation:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/ResultList'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    ResultList:
      type: object
      properties:
        total_hits:
          type: integer
        total_pages:
          type: integer
        current_page:
          type: integer
        suggestion:
          type: string
        search_time:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/Result'
    Result:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
        abs:
          type: string
        highlight:
          type: string
        reading_time:
          type: integer
        date_creation:
          type: string
        isFavorite:
          type: boolean
    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: A human readable error message
          type: string